[Refactor] #45 섹션 분리
