+++ src/components/Main/WeatherSection.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import SectionCard from './SectionCard';
+
+const WeatherSection = () => {
+    return (
+        <SectionCard title="오늘날씨">
+            {/* 날씨 컨텐츠 */}
+        </SectionCard>
+    );
+};
+
+export default WeatherSection;
