+++ src/components/Main/WorkoutTrendSection.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const WorkoutTrendSection = () => {
+    return (
+        <aside className="md:w-4/12 flex-2 bg-mainRed rounded-md opacity-60 p-4">
+            <header>
+                <h2 className="text-lg text-mainText ml-2">이번주 운동량 추세</h2>
+            </header>
+            {/* 운동량 추세 컨텐츠 */}
+        </aside>
+    );
+};
+
+export default WorkoutTrendSection;
